Create package, support web_app target
