+++ environment/configuration/requirements.txt
@@ -2,4 +2,3 @@ docopt
 matplotlib
 networkx[default]
 pyside6>=6.6
-flet==0.22.*
