+++ source/CMakeLists.txt
@@ -1,2 +1,3 @@
 add_subdirectory(package/adaptation_pathways/desktop)
+add_subdirectory(package/pathways_app)
 add_subdirectory(test)
