+++ source/package/pathways_app/README.md
@@ -3,11 +3,21 @@
 To run the app from within this directory:
 
 ```
-flet run .
+flet run main.py
 ```
 
 To run the app as a website:
 
 ```
-flet run --web .
+flet run --web
+```
+
+To build the web app:
+
+```
+# Assumes you're in the binary directory
+cmake --build . --target web_app
+
+# Alternatively, if using Ninja for example
+ninja web_app
 ```
