+++ source/package/pathways_app/pathways_app/controls/actions_panel.py
@@ -2,18 +2,19 @@
 import random
 
 import flet as ft
-import theme
-from controls.action_icon import ActionIcon
-from controls.editable_cell import EditableTextCell
-from controls.metric_effect import MetricEffectCell
-from controls.metric_value import MetricValueCell
-from controls.sortable_header import SortableHeader, SortMode
-from controls.styled_button import StyledButton
-from controls.styled_table import StyledTable
 
 from adaptation_pathways.app.model.action import Action
 from adaptation_pathways.app.model.pathways_project import PathwaysProject
 from adaptation_pathways.app.model.sorting import SortTarget
+
+from .. import theme
+from .action_icon import ActionIcon
+from .editable_cell import EditableTextCell
+from .metric_effect import MetricEffectCell
+from .metric_value import MetricValueCell
+from .sortable_header import SortableHeader, SortMode
+from .styled_button import StyledButton
+from .styled_table import StyledTable
 
 
 class ActionsPanel(ft.Column):
