+++ source/package/pathways_app/pathways_app/controls/header.py
@@ -1,6 +1,7 @@
 # pylint: disable=too-many-arguments
 import flet as ft
-import theme
+
+from .. import theme
 
 
 class SectionHeader(ft.Container):
