+++ source/package/pathways_app/pathways_app/controls/menu_bar.py
@@ -1,7 +1,8 @@
 import flet as ft
-import theme
 
 from adaptation_pathways.app.model import PathwaysProject
+
+from .. import theme
 
 
 class MenuBar(ft.Container):
