+++ source/package/pathways_app/pathways_app/controls/panel.py
@@ -1,5 +1,6 @@
 import flet as ft
-import theme
+
+from .. import theme
 
 
 class Panel(ft.Container):
