+++ source/package/pathways_app/pathways_app/controls/styled_table.py
@@ -1,6 +1,7 @@
 # pylint: disable=too-many-arguments
 import flet as ft
-import theme
+
+from .. import theme
 
 
 class StyledTable(ft.DataTable):
