+++ source/package/pathways_app/pathways_app/controls/tabbed_panel.py
@@ -1,6 +1,7 @@
 import flet as ft
-import theme
-from controls.header import SectionHeader
+
+from .. import theme
+from .header import SectionHeader
 
 
 class TabbedPanel(ft.Column):
