+++ source/package/pathways_app/pathways_app/controls/unit_cell.py
@@ -1,9 +1,10 @@
 from typing import Callable
 
 import flet as ft
-import theme
 
 from adaptation_pathways.app.model.metric import Metric, MetricUnit, default_units
+
+from .. import theme
 
 
 class MetricUnitCell(ft.DataCell):
