--- source/package/pathways_app/requirements.txt
@@ -1,3 +1,0 @@
-matplotlib
-pyparsing
-networkx
